pyproject.toml
[project]
name = "email-assistant"
version = "0.1.0"
description = "CLI Gmail assistant that auto-labels incoming messages"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.1",
    "google-api-python-client>=2.190.0",
    "google-auth>=2.48.0",
    "google-auth-oauthlib>=1.3.0",
    "python-dotenv>=1.2.1",
    "rich>=14.3.3",
]

[project.optional-dependencies]
ml = [
    "joblib>=1.5.3",
    "scikit-learn>=1.8.0",
]

[project.scripts]
email-assistant = "main:cli"
